fix varie

+++ public/js/pages.js
@@ -89,24 +89,30 @@ $(document).ready(function() {
     // Get the form.
 
     $(".delete-button").on("click",function(){
+
       event.preventDefault();
       event.stopImmediatePropagation();
 
-      var request = $.ajax({
-        type: 'DELETE',
-        url: "/api/page/"+$(this).attr('id')
-      });
+      if (confirm('Sei sicuro di voler cancellare la tua pagina? Una volta cliccato OK non sarà possibile tornare indietro!')) {
+        var request = $.ajax({
+          type: 'DELETE',
+          url: "/api/page/"+$(this).attr('id')
+        });
 
-      request.done(function(response) {
-        // Make sure that the formMessages div has the 'success' class.
-        alert("Informazioni aggiornate!");
-        window.location = '/console/admin-console/pages';
+        request.done(function(response) {
+          // Make sure that the formMessages div has the 'success' class.
+          alert("Pagina cancellata!");
+          window.location = '/console/admin-console/pages';
 
-      });
-      request.fail(function(data) {
-        // Make sure that the formMessages div has the 'error' class.
-        alert("Qualcosa è andato storto! Riprova!");
-      });
+        });
+        request.fail(function(data) {
+          // Make sure that the formMessages div has the 'error' class.
+          alert("Qualcosa è andato storto! Riprova!");
+        });
+
+      } else {
+        alert('Sei stato saggio');
+      }
 
     });
